WIP: saving state from end of 3/29/2019 puttering
We are/were trying to figure out how to write data from memory to a GPIO
port with DMA1. We started with TIM12_TRGO generating DMA Request 7 but
ran into trouble so switched our approach to try using HRTIM to generate
the request instead. However, we had trouble getting Timer A,
output 1 to show up on the scope (via PC6).